vdx_id_agent/requirements.txt
backoff==1.9.2
bson==0.5.10
celery[redis]==4.4.7
circus==0.15.0
cryptography==3.3.2
deepdiff==4.0.9
environs==6.1.0
fabric==2.5.0
pymongo[tls]==3.9.0
requests==2.22.0
watchdog==0.9.0
numpy==1.18.1
pandas==0.25.3
pycryptodome==3.10.1
psycopg2==2.8.5
python-ldap==3.3.0
redis==3.3.11
cx_Oracle==8.1.0
rejson==0.5.4
